Extract bird
