+++ src/Pages/Home.svelte
@@ -1,6 +1,7 @@
 <script lang="ts">
     import gsap from "gsap";
     import { onMount } from "svelte";
+    import Bird from "../Components/Bird.svelte";
     import RatingFiller from "../Components/Ratings/RatingFiller.svelte";
 
     gsap.registerPlugin(ScrollTrigger);
@@ -13,37 +14,13 @@
 
     console.log("Inner Height = " + window.innerHeight);
     console.log("Inner Width = " + window.innerWidth);
-
-    onMount(() => {
-        gsap.fromTo(
-            "#bird",
-            { opacity: 1 },
-            {
-                y: -250,
-                x: 800,
-                ease: "power2.out",
-                scrollTrigger: {
-                    trigger: ".scrollElement",
-                    start: "0% top",
-                    end: "10% top",
-                    scrub: 1,
-                    onEnter: function () {
-                        gsap.to("#bird", { scaleX: 1, rotation: 0 });
-                    },
-                    onLeave: function () {
-                        gsap.to("#bird", { scaleX: -1, rotation: -15 });
-                    },
-                },
-            }
-        );
-    });
 </script>
 
 <RatingFiller />
 <div style="min-height: 1000vh;">
     <div class="rows">
         <h1 bind:this={title} style="font-family: earthsmightiestbold; color: var(--COLOR_PRIMARY); ">Tceliana</h1>
-        <p id="bird">BIRD</p>
+        <Bird />
         <p>The story begins.</p>
         <img class="hearth" src="/Images/hearth.svg" alt="hearth SVG" />
     </div>
